% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2ReasyR_functions.R
\name{add.dot.bracket}
\alias{add.dot.bracket}
\title{Converts RNA secondary structure information in a .ct file into dot-bracket notation}
\usage{
add.dot.bracket(ctdata_frame)
}
\arguments{
\item{ctdata_frame}{A data frame made by reading a .ct file into R with read.ct}
}
\value{
A data frame containing a collumn with dot-bracket formated RNA secondary structure
}
\description{
This function reformats information in a dataframe, made from reading a .ct file into R,
into the dot-bracket notation.
}
