% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.accn.R
\name{check.accn}
\alias{check.accn}
\title{Makes a list of accension numbers that are the same}
\usage{
check.accn(data1, data2)
}
\arguments{
\item{data1}{The first data frame you want to cross reference}

\item{data2}{The second data frame you want to cross reference}
}
\value{
A data frame containing rows in data1 that contain accension numbers that are also in data2
}
\description{
This function cross references the accenssion numbers in two data
frames to determine which accenssion numbers are in both data frames. It
requires that both data frames have a column titled "Accension", with the
Acension numbers that you want to cross reference.
}
