% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peter.ct.reformat.R
\name{peter.ct.reformat}
\alias{peter.ct.reformat}
\title{A function that reformats Peter's double tabdelimited files for reading in to the R2easyR pipeline}
\usage{
peter.ct.reformat(file.ct)
}
\arguments{
\item{file.ct}{Path to the ct file that you want to reformat}
}
\description{
This function edits a double tab delimited ct (conectivity file) to make it compatible with the read.ct function in R2easyR.
}
