% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2ReasyR_functions.R
\name{r2easyR.custom.palette}
\alias{r2easyR.custom.palette}
\title{Generates a custom palette}
\usage{
r2easyR.custom.palette(colors)
}
\arguments{
\item{colors}{A vector containing less than 35 R colors}
}
\value{
A custom palette that can be used by r2easyR.color
}
\description{
Makes a custom R2Reasy palette from a vector containing less than 35 colors
}
