% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2ReasyR_functions.R
\name{r2easyR.custom.pallet}
\alias{r2easyR.custom.pallet}
\title{Generates a custom pallet}
\usage{
r2easyR.custom.pallet(colors)
}
\arguments{
\item{colors}{A vector containing less than 35 R colors}
}
\value{
A custom pallet that can be used by r2easyR.color
}
\description{
Makes a custom R2Reasy pallet from a vector containing less than 35 colors
}
