% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2easyR.go_fast.R
\name{r2easyR.go_fast}
\alias{r2easyR.go_fast}
\title{Prodeces R2R inputs in bulk with optimized layouts}
\usage{
r2easyR.go_fast(
  CT_folder = "CT",
  Reactivity_folder = "Reactivity",
  R2R_folder = "R2R_files",
  abs_reactivity_threshold = 0.01,
  manual.scale = c(0, 2),
  RNA_name = "RNA",
  RNA_to_make_legend_on = 1
)
}
\arguments{
\item{CT_folder}{Path to the directory containing .ct files}

\item{Reactivity_folder}{Path to the directory containing .shape files}

\item{R2R_folder}{Path to the directory you want the R2R .sto and metafiles to print in}

\item{abs_reactivity_threshold}{Minimum threshold for reactivity data to be mapped to a color. Default = 0. Using a higher threshold prevents numerous, but low reactivity values from cluttering up the finished picture.}

\item{manual.scale}{Manually specify a min and max. Useful if you want to color reactivity from multiple transcripts using 1 scale. Default is "FALSE". To set the scale, use manual.scale = c(min, max)}

\item{RNA_name}{The name of the RNA that you are drawing}

\item{RNA_to_make_legend_on}{You will need a legend to interperet the results. This program will print one for you. It will also print a reaction plot for you to see what the pallets look like on data. Set this to determine which ".ct" file you want to print the reaction plot and legend on (Legends will be the same for every structure because all structures use the sane manual scale)}
}
\description{
This function reads a folder of RNAstructure CT files and a folder of ".shape" formated reactivity data and produces
R2R inputs with optimized stem layouts in bulk. Enables you to make lots of RNA secondary structures with two lines of code.
}
