% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2easyR.grey_letters_editor.R
\name{r2easyR.grey_letters_editor}
\alias{r2easyR.grey_letters_editor}
\title{This function edits a R2R stockholm file to recolor nucleotides in a secondary structure.}
\usage{
r2easyR.grey_letters_editor(R2R.sto, Nucleotides, Color = "dimgrey")
}
\arguments{
\item{R2R.sto}{Path to the R2R stockholm file that contains the drawing information you want to optomize}

\item{Nucleotides}{A vector listing the nucleotide numbers you want to color}

\item{Color}{The color you want the nucleotides to be. Use R color formats. Default is Dim grey}
}
\description{
This function edits a R2R stockholm file to recolor nucleotides in a secondary structure.
}
