% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2ReasyR_functions.R
\name{r2easyR.palettes}
\alias{r2easyR.palettes}
\title{Generates a list of palettes}
\usage{
r2easyR.palettes()
}
\value{
A list of 59 vectors containing Viridis and Colorbrewer palettes
}
\description{
Makes a list of 59 color palettes for r2easyR.colors. Also saves a PDF depicting
all of the color palettes in your current working directory. Color palettes are generated using RColorBrewer
or Viridis. ".l" palettes are recommended for coloring letters because they exclude light shades, which would be
hard to see against a white background. ".c" palettes are recommended for coloring circles behind letters
because they exclude dark colors, which obscure the letter inside the circle. Viridis palettes, Viridis, Magma,
Plasma, Inferno, and cividis are not recommended because they result in cluttered secondary structures.
}
