% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2ReasyR_functions.R
\name{r2easyR.pallets}
\alias{r2easyR.pallets}
\title{Generates a list of pallets}
\usage{
r2easyR.pallets()
}
\value{
A list of 32 vectors containing Viridis and Colorbrewer pallets
}
\description{
Makes a list of 59 color pallets for r2easyR.colors. Also saves a PDF depicting
all of the color pallets in your current working directory. Color pallets are generated using RColorBrewer
or Viridis. ".l" pallets are recomended for coloring letters because they exclude light shades, which would be
hard to see against a white background. ".c" pallets are recomended for for coloring circles behing letters
because they exclude dark colors, which obscure the letter inside the circle. Viridis pallets, Viridis, Magma,
Plasma, Inferno, and cividis are not recomended because they result in cluttered secondary structures.
}
