% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2easyR.pk_finder.R
\name{r2easyR.pk_finder}
\alias{r2easyR.pk_finder}
\title{Identifies non-nested base pairs (pseudoknots) from an RNA secondary structure specified
by a connectivity table (.ct) formatted data frame.}
\usage{
r2easyR.pk_finder(ctdata_frame)
}
\arguments{
\item{ctdata_frame}{A data frame made by reading a .ct file into R with read.ct}
}
\value{
A list of two elements, the first is the modified data frame with the pseudoknots deleted from the dotbracket column and the second is a list of vectors identifying pseudonotted bases.
}
\description{
Capable of identifying multiple pseudoknots and pseudoknots within pseudoknots.
Returns a modified data frame with pseudoknotted base pairs deleted and a list
of vectors of the pseudoknot base pairs. Pseudoknots can be passed to the r2easyR.pknot_drawer.
Pseudoknots are troublesome for structure drawing programs because their non-nested nature
fools the drawing program into pairing the incorrect bases.
}
