% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2easyR.pknot_drawer.R
\name{r2easyR.pknot_drawer}
\alias{r2easyR.pknot_drawer}
\title{Edits a R2R stockholm file for drawing psuedoknot labels on RNA secondary structures}
\usage{
r2easyR.pknot_drawer(R2R.sto, pknot)
}
\arguments{
\item{R2R.sto}{Path to the R2R stockholm file that contains the drawing information you want to add a pknot to}

\item{pknot}{A vector containing pknot nucleotide numbers}
}
\description{
Useful for adding psuedoknot labels to RNA secondary structures drawn with R2R. Counts the labels already specified by the stockholm
and adjusts accordingly.
}
