% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2easyR.tRNA_stem_editor.R
\name{r2easyR.tRNA_stem_editor}
\alias{r2easyR.tRNA_stem_editor}
\title{A function that optomizes the stem layout for a tRNA 2-structure drawn with R2R}
\usage{
r2easyR.tRNA_stem_editor(R2R.sto)
}
\arguments{
\item{R2R.sto}{Path to the R2R stockholm file that contains the drawing information you want to optomize}
}
\description{
This function edits a R2R stockholm file to provide a better stem layout for tRNA.
}
