% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2ReasyR_functions.R
\name{r2easyR.write}
\alias{r2easyR.write}
\title{Generates a Stockholm file and a R2R meta file that can be read by R2R to draw a secondary structure.}
\usage{
r2easyR.write(output, data_frame, RNA_name = "default", colors = "NA")
}
\arguments{
\item{output}{Path to the output Stockholm file}

\item{data_frame}{A data_frame containing columns labeled "Nucleotide", "Dotbracket", "Labels", and "Colors".}

\item{RNA_name}{The name of the RNA that you are drawing}

\item{colors}{How R2R will draw reactivity colors. No colors = "NA". Colored letters = "letters". Colored circles = "circles".}
}
\value{
A Stockholm formated file.
}
\description{
Generates a Stockholm file and a R2R meta file that can be read by R2R to draw a secondary structure.
}
