% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.connect.R
\name{read.connect}
\alias{read.connect}
\title{Read a .connect file file}
\usage{
read.connect(data_file, T.to.U = TRUE)
}
\arguments{
\item{data_file}{Path to the data file}

\item{T.to.U}{Default = TRUE then read connect will convert all "T"s to "U"s. If false read.connect will leave "T"s.}
}
\value{
A data frame containing the information in the .ct file
}
\description{
This function reads a .ct file, from the CoFold RNA folding algorithm, into a R data frame.
}
