% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2ReasyR_functions.R
\name{read.ct}
\alias{read.ct}
\title{Read a .ct file}
\usage{
read.ct(data_file)
}
\arguments{
\item{data_file}{Path to the data file}
}
\value{
A data frame containing the information in the .ct file
}
\description{
This function reads a .ct file, from RNA folding algorithms, into a R data frame.
}
