% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.genbank.csv.R
\name{read.genbank.csv}
\alias{read.genbank.csv}
\title{Read a csv file from genbank}
\usage{
read.genbank.csv(csv_file)
}
\arguments{
\item{csv_file}{Path to the csv formated data file in GENBANK}
}
\value{
A data frame containing the information in the csv formated data file
}
\description{
This function reads a csv formated allignment from the GENBANK database into a R
data.frame. It also separates the Accension number from other data into
separate column in the data frame. Data in data.frames are stored as strings.
}
