% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.rnabob.R
\name{read.rnabob}
\alias{read.rnabob}
\title{Loads an RNABOB output file into a R data frame}
\usage{
read.rnabob(data_file)
}
\arguments{
\item{data_file}{Path to the RNABOB output file}
}
\value{
A data frame containing the data from the RNABOB output file
}
\description{
This function reads a RNABOBoutput file into a R data frame
}
