% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.rtsc.R
\name{read.rtsc}
\alias{read.rtsc}
\title{Read a .rtsc file}
\usage{
read.rtsc(data_file)
}
\arguments{
\item{data_file}{Path to the data file}
}
\value{
A list of vectors containing stop count data from the different RNA in the .rtsc file
}
\description{
This function reads a .rtsc file from StructureFold2 into R
}
