% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2ReasyR_functions.R
\name{read.shape}
\alias{read.shape}
\title{Read a .shape file}
\usage{
read.shape(data_file)
}
\arguments{
\item{data_file}{Path to the data file}
}
\value{
A list of vectors containing reactivity data from the different RNA in the .react file
}
\description{
This function reads a .react file from StructureFold2 into R
}
