% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.stockholm.R
\name{read.stock}
\alias{read.stock}
\title{Read a Stockholm file}
\usage{
read.stock(data_file)
}
\arguments{
\item{data_file}{Path to the data file}
}
\value{
A data frame containing the information in the Stockholm file
}
\description{
This function reads a Stockholm formated allignment from the RFAM database into a R
data.frame. It trims the long header at the front of the Stockholm file and keeps the
data. It also separates the Accension number and the sequence location into
separate column in the data frame. Data in data.frames are stored as strings.
}
