% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFAM_sto_to_ct.R
\name{RFAM_sto_to_ct}
\alias{RFAM_sto_to_ct}
\title{A that converts a RFAM formatted sto file to a connectivity (CT) table file}
\usage{
RFAM_sto_to_ct(
  Path.to.sto,
  RNA.name,
  Output.name = "RNA",
  Remove.non.canonical = TRUE,
  Keep.pknots = TRUE
)
}
\arguments{
\item{Path.to.sto}{Path to the sto file that you want to reformat}

\item{RNA.name}{Name of the RNA that you want to pull out}

\item{Output.name}{Prefix you want to put on the output CT file. Default = "RNA". Can be a path to a CT file storage directory.}

\item{Remove.non.canonical}{Do you want to remove non canonical and non GU wobble pairs produced from the consensus allignment. Default = TRUE. Set to Remove.non.canonical = FALSE.}

\item{Keep.pknots}{Do you want to keep pknots? Default = TRUE. Set to Keep.pknots = FALSE to remove pknots.}
}
\description{
Parses a RFAM formatted Stockholm alignment file, pulls out the sequence you want, and maps the consensus secondary structure onto a
connectivity table (CT).
}
