% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what.bob.knows.R
\name{What.bob.knows}
\alias{What.bob.knows}
\title{Identifies what known RNA were identified by a motif seach using RNAbob}
\usage{
What.bob.knows(
  known.Accenssion,
  known.Sequence,
  hit.Accenssion,
  hit.Sequence,
  hit.Descriptor
)
}
\arguments{
\item{known.Accenssion}{Vector containing known accenssion numbers}

\item{known.Sequence}{Vector containing known RNA sequences}

\item{hit.Accenssion}{Vector containing hit accension numbers}

\item{hit.Sequence}{Vector containing hit sequences}

\item{hit.Descriptor}{Vector containing hit descriptors}
}
\value{
A R data frame containing the number of times RNAbob found a known RNA, the descriptors that found each RNA, and the RNA sequences that RNAbob found
}
\description{
Identifies what known RNA were identified by a motif seach using RNAbob
}
